Seating Diagrams E2 and E6
Both seating diagrams can be used to capture any room.  Each table element will go blank if an element name is null so we just need the database to reflect null values for undetected spaces.  The JS files just need to be tied to the variables passed by the database to them and it will work perfectly.